requirements.txt
ag2==0.8.6
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
elastic-transport==8.17.1
elasticsearch==8.17.2
exceptiongroup==1.2.2
fastapi>=0.68.0
ffmpy==0.5.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
gliner==0.2.17
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx>=0.18.0
huggingface-hub==0.30.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_resources==6.5.2
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
onnxruntime==1.19.2
openai==1.74.0
orjson==3.10.16
owlready2==0.47
packaging==24.2
pandas>=1.3.0
pillow==10.4.0
portalocker==2.10.1
protobuf==5.29.4
pyautogen==0.8.6
pydantic>=1.8.0
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv>=0.19.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.13.3
rdflib>=6.0.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruff==0.11.5
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==4.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SPARQLWrapper>=2.0.0
starlette==0.46.2
sympy==1.13.1
termcolor==3.0.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.12.0
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn>=0.15.0
websockets==12.0
zipp==3.21.0
redis>=4.0.0
celery>=5.2.0
dask>=2023.0.0
ray>=2.0.0
confluent-kafka>=2.0.0
python-jose>=3.3.0
